refactor!: change caesar_cipher.rb for tests
1. Change file structure and set up tests

BREAKING CHANGE: Files with the original code in are now in the lib
directory

2. Get caesar_cipher from script to class

This allows us to test it as a class in the usual way.

BREAKING CHANGE: The file now cannot be run independently as a script.
The object needs to be instantiated and the play method called.

3. Split up input into smaller methods

This makes this easier to test.

BREAKING CHANGE: Some public methods have changed for input.

4. Change verified_shift to looping script

To make it more easily identifiable for testing.

5. Clean up apply_shift

Reduced lines of code and made shift_char private as it is an
implementation detail mostly involving maths.

BREAKING CHANGE: shift_char is now private

6. Change initialize to take input message

This allows for easier testing.

The user input to update message is now contained in the play script
method which also makes it easier to understand as the process is in one
place.

Put logic to exit the loop in verified_shift back into the method itself
since separating this out only adds to the number of tests without need.

BREAKING CHANGE: initialize now takes an argument instead of user input.
play() now acts differently. The method for the logic to exit the loop
in verified_shift is now gone.